requirements.txt
pymupdf4llm
rapidfuzz>=0.0.1
pymupdf>=1.24.0

# Mendeley API (v2 script uses requests directly, not the deprecated SDK)
requests>=2.32.0
python-dotenv
